Move direction namespace to another module
